moved ShellSort, added README for searching and sorting
