+++ programs/sorting/README.md
@@ -1,0 +1,43 @@
+## 3 Searching
+
+
+| Program              | Data files    |
+| :--------            |:-------------:|
+| CompareSort          | * none, but takes two algorithms and two numbers as arguments * |
+| JavaArraySort        | sorts every *word/number* it reads from standard input. You can therefore use `&lt;` and literally any text file. *algs4-data/&#42;Kints.txt* is a good place to start. |
+| MySort               | * none * |
+| SortCompare          | * same as CompareSort * |
+
+
+### Elementary Sorts
+
+| Program              | Data files    |
+| :--------            |:-------------:|
+| ExampleSort          | * same as JavaArraySort * |
+| CountSort99          | * none, but takes a number as argument * |
+| InsertionSort        | * same as JavaArraySort * |
+| ShellSort            | * same as JavaArraySort * |
+
+### mergeSort
+
+| Program              | Data files    |
+| :--------            |:-------------:|
+| BottomUpMergeSort    | * same as JavaArraySort * |
+| TopDownMergeSort     | * same as JavaArraySort * |
+
+
+### priorityQueues
+
+| Program              | Data files    |
+| :--------            |:-------------:|
+| ArrayListPQ          | * none, but takes a number as argument * |
+| BottomM              | * none, but takes a number as argument * |
+| MultiwayMerge        | * any type of text file, as many files as you want. Takes filenames as arguments * |
+| NaiveHeapSort        | * same as JavaArraySort * |
+
+
+### quickSort
+
+| Program              | Data files    |
+| :--------            |:-------------:|
+| QuickSort            | * same as JavaArraySort * |
