+++ programs/sorting/elementarySorts/ShellSort.java
@@ -1,4 +1,4 @@
-package sorting;
+package sorting.elementarySorts;
 
 import edu.princeton.cs.algs4.In; import edu.princeton.cs.algs4.StdOut;
 public class ShellSort {
